requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis[hiredis]>=5.0.0
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
